drop_caches: add some documentation and info message
There is plenty of anecdotal evidence and a load of blog posts suggesting
that using "drop_caches" periodically keeps your system running in "tip
top shape".  Perhaps adding some kernel documentation will increase the
amount of accurate data on its use.

If we are not shrinking caches effectively, then we have real bugs.  Using
drop_caches will simply mask the bugs and make them harder to find, but
certainly does not fix them, nor is it an appropriate "workaround" to
limit the size of the caches.  On the contrary, there have been bug
reports on issues that turned out to be misguided use of cache dropping.

Dropping caches is a very drastic and disruptive operation that is good
for debugging and running tests, but if it creates bug reports from
production use, kernel developers should be aware of its use.

Add a bit more documentation about it, a syslog message to track down
abusers, and vmstat drop counters to help analyze problem reports.

[akpm@linux-foundation.org: checkpatch fixes]
[hannes@cmpxchg.org: add runtime suppression control]
Signed-off-by: Dave Hansen <dave@linux.vnet.ibm.com>
Signed-off-by: Michal Hocko <mhocko@suse.cz>
Acked-by: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
Acked-by: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>