sched,numa: add cond_resched to task_numa_work
Normally task_numa_work scans over a fairly small amount of memory, but it
is possible to run into a large unpopulated part of virtual memory, with
no pages mapped.  In that case, task_numa_work can run for a while, and it
may make sense to reschedule as required.

Signed-off-by: Rik van Riel <riel@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Reported-by: Xing Gang <gang.xing@hp.com>
Tested-by: Chegu Vinod <chegu_vinod@hp.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>